Extract Built and Compiled releases into own files
Authored-by: Matt Royal <mroyal@pivotal.io>
