+++ fetcher/built_release.go
@@ -1,0 +1,10 @@
+package fetcher
+
+type BuiltRelease struct {
+	ID   ReleaseID
+	Path string
+}
+
+func (br BuiltRelease) DownloadString() string {
+	return br.Path
+}
